+++ packages/pn-commons/src/components/NotificationDetail/NotificationDetailTimelineStep.tsx
@@ -281,24 +281,22 @@ const NotificationDetailTimelineStep = ({
             </Typography>
             <Box sx={{ overflowWrap: 'anywhere' }}>
               <Typography color="text.primary" fontSize={14}>
-                {timelineStatusInfos.description}&nbsp;
-                {s.legalFactsIds &&
-                  s.legalFactsIds.length > 0 &&
-                  s.legalFactsIds.map((lf) => (
-                    <Typography
-                      fontSize={14}
-                      display="inline"
-                      variant="button"
-                      color={disableDownloads ? 'text.disabled' : 'primary'}
-                      sx={{ cursor: disableDownloads ? 'default' : 'pointer' }}
-                      onClick={() => clickHandler(lf)}
-                      key={lf.key}
-                      data-testid="download-legalfact-micro"
-                    >
-                      {getLegalFactLabel(s, lf.category, lf.key || '')}
-                    </Typography>
-                  ))}
+                {timelineStatusInfos.description}
               </Typography>
+              {s.legalFactsIds &&
+                s.legalFactsIds.length > 0 &&
+                s.legalFactsIds.map((lf) => (
+                  <ButtonNaked
+                    fontSize={14}
+                    color='primary'
+                    onClick={() => clickHandler(lf)}
+                    disabled={disableDownloads}
+                    key={lf.key}
+                    data-testid="download-legalfact-micro"
+                  >
+                    {getLegalFactLabel(s, lf.category, lf.key || '')}
+                  </ButtonNaked>
+                ))}
             </Box>
           </Fragment>
         }