+++ packages/pn-pa-webapp/src/components/Notifications/DesktopNotifications.tsx
@@ -1,6 +1,6 @@
 import { useRef } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
-import { useNavigate } from 'react-router-dom';
+ import { useNavigate } from 'react-router-dom';
 
 import { Link } from '@mui/material';
 import {
@@ -14,10 +14,11 @@ import {
   PnTableBodyRow,
   PnTableHeader,
   PnTableHeaderCell,
-  Row,
+  // Row,
   Sort,
 } from '@pagopa-pn/pn-commons';
 
+import { NotificationColumnData } from '@pagopa-pn/pn-commons/src/models/Notifications';
 import * as routes from '../../navigation/routes.const';
 import FilterNotifications from './FilterNotifications';
 import NotificationsDataSwitch from './NotificationsDataSwitch';
@@ -25,9 +26,9 @@ import NotificationsDataSwitch from './NotificationsDataSwitch';
 type Props = {
   notifications: Array<Notification>;
   /** Table sort */
-  sort?: Sort<Notification>;
+  sort?: Sort<NotificationColumnData>;
   /** The function to be invoked if the user change sorting */
-  onChangeSorting?: (s: Sort<Notification>) => void;
+  onChangeSorting?: (s: Sort<NotificationColumnData>) => void;
   /** The function to be invoked if the user clicks on new notification link */
   onManualSend: () => void;
   /** The function to be invoked if the user clicks on api keys link */
@@ -98,11 +99,11 @@ const DesktopNotifications = ({
   onManualSend,
   onApiKeys,
 }: Props) => {
-  const navigate = useNavigate();
   const filterNotificationsRef = useRef({ filtersApplied: false, cleanFilters: () => void 0 });
   const { t } = useTranslation(['notifiche']);
+  const navigate = useNavigate();
 
-  const columns: Array<Column<Notification>> = [
+  const columns: Array<Column<NotificationColumnData>> = [
     {
       id: 'sentAt',
       label: t('table.date'),
@@ -118,7 +119,7 @@ const DesktopNotifications = ({
     {
       id: 'subject',
       label: t('table.subject'),
-      cellProps: { width: '23%' },
+      cellProps: { width: '18%' },
     },
     {
       id: 'iun',
@@ -136,6 +137,12 @@ const DesktopNotifications = ({
       cellProps: { width: '18%' },
       sortable: false, // TODO: will be re-enabled in PN-1124
     },
+    {
+      id: 'action',
+      label: '',
+      cellProps: { width: '8%' },
+      sortable: false,
+    },
   ];
 
   const rows = notifications.map((n) => ({
@@ -143,14 +150,14 @@ const DesktopNotifications = ({
     id: n.iun,
   }));
 
-  // Navigation handlers
-  const handleRowClick = (row: Row<Notification>) => {
-    navigate(routes.GET_DETTAGLIO_NOTIFICA_PATH(row.iun));
-  };
-
   const filtersApplied: boolean = filterNotificationsRef.current.filtersApplied;
 
   const showFilters = notifications?.length > 0 || filtersApplied;
+
+    // Navigation handlers
+    const handleRowClick = (iun: string) => {
+      navigate(routes.GET_DETTAGLIO_NOTIFICA_PATH(iun));
+    };
 
   return (
     <>
@@ -178,13 +185,13 @@ const DesktopNotifications = ({
                     {columns.map((column) => (
                       <PnTableBodyCell
                         key={column.id}
-                        onClick={() => handleRowClick(row)}
+                        // onClick={() => handleRowClick(row)}
                         cellProps={{
                           ...column.cellProps,
-                          cursor: 'pointer',
+                          // cursor: 'pointer',
                         }}
                       >
-                        <NotificationsDataSwitch data={row} type={column.id} />
+                        <NotificationsDataSwitch handleRowClick={handleRowClick} data={row} type={column.id} />
                       </PnTableBodyCell>
                     ))}
                   </PnTableBodyRow>