+++ packages/pn-personagiuridica-webapp/src/components/Notifications/DesktopNotifications.tsx
@@ -99,6 +99,11 @@ const DesktopNotifications = ({
       cellProps: { width: '18%' },
       sortable: false, // TODO: will be re-enabled in PN-1124
     },
+    {
+      id: 'action',
+      label: '',
+      cellProps: { width: '8%' },
+    },
   ];
 
   if (isDelegatedPage) {
@@ -122,11 +127,11 @@ const DesktopNotifications = ({
   const showFilters = notifications?.length > 0 || filtersApplied;
 
   // Navigation handlers
-  const handleRowClick = (row: Row<Notification>) => {
-    if (isDelegatedPage && row.mandateId) {
-      navigate(routes.GET_DETTAGLIO_NOTIFICA_DELEGATO_PATH(row.iun, row.mandateId));
+  const handleRowClick = (iun: string ,mandateId?: string) => {
+    if (isDelegatedPage && mandateId) {
+      navigate(routes.GET_DETTAGLIO_NOTIFICA_DELEGATO_PATH(iun, mandateId));
     } else {
-      navigate(routes.GET_DETTAGLIO_NOTIFICA_PATH(row.iun));
+      navigate(routes.GET_DETTAGLIO_NOTIFICA_PATH(iun));
     }
   };
 
@@ -154,13 +159,13 @@ const DesktopNotifications = ({
                 {columns.map((column) => (
                   <PnTableBodyCell
                     key={column.id}
-                    onClick={() => handleRowClick(row)}
+                    // onClick={() => handleRowClick(row)}
                     cellProps={{
                       ...column.cellProps,
-                      cursor: 'pointer',
+                      // cursor: 'pointer',
                     }}
                   >
-                    <NotificationsDataSwitch data={row} type={column.id} />
+                    <NotificationsDataSwitch handleRowClick={handleRowClick} data={row} type={column.id} />
                   </PnTableBodyCell>
                 ))}
               </PnTableBodyRow>